+++ Kexa/models/export/config.models.ts
@@ -1,0 +1,9 @@
+export interface SaveConfig {
+    type: string;
+    urlName: string;
+    name?: string;
+    description?: string;
+    origin?: string;
+    tags?: {[key: string]: string};
+    onlyErrors?: boolean;
+}