+++ Kexa/models/export/mongoDB/config.models.ts
@@ -1,0 +1,6 @@
+import { SaveConfig } from "../config.models";
+
+export interface MongoDBSaveConfig extends SaveConfig {
+    type: "mongoDB";
+    collectionName: string;
+}