+++ Kexa/models/settingFile/config.models.ts
@@ -1,5 +1,6 @@
 import { AwsConfig } from "../aws/config.models";
 import { AzureConfig } from "../azure/config.models";
+import { SaveConfig } from "../export/config.models";
 import { GcpConfig } from "../gcp/config.models";
 import { GitConfig } from "../git/config.models";
 import { HttpConfig } from "../http/config.models";
@@ -12,6 +13,7 @@ export interface GlobalConfig{
     kubernetes?: KubernetesConfig[];
     git?: GitConfig[];
     http?: HttpConfig[];
+    save?: SaveConfig[];
     EMAILPORT?: string;
     EMAILHOST?: string;
     EMAILUSER?: string;